refactor(API): expose resource types to plugins
