+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/resolvers/KeyPairResolver.kt
@@ -2,10 +2,10 @@ package com.netflix.spinnaker.keel.ec2.resolvers
 
 import com.netflix.spinnaker.keel.api.Resource
 import com.netflix.spinnaker.keel.api.ec2.ClusterSpec
+import com.netflix.spinnaker.keel.api.ec2.EC2_CLUSTER_V1
 import com.netflix.spinnaker.keel.api.ec2.LaunchConfigurationSpec
 import com.netflix.spinnaker.keel.api.plugins.Resolver
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverCache
-import com.netflix.spinnaker.keel.ec2.EC2_CLUSTER_V1
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
